+++ src/templatesCompiled/client/zod/partials/zodSchemaArray.ts
@@ -12,8 +12,8 @@ export default {"1":function(container,depth0,helpers,partials,data) {
         return undefined
     };
 
-  return ((stack1 = lookupProperty(helpers,"if").call(depth0 != null ? depth0 : (container.nullContext || {}),lookupProperty(lookupProperty(depth0,"link"),"alias"),{"name":"if","hash":{},"fn":container.program(2, data, 0),"inverse":container.program(4, data, 0),"data":data,"loc":{"start":{"line":1,"column":20},"end":{"line":1,"column":96}}})) != null ? stack1 : "");
-},"2":function(container,depth0,helpers,partials,data) {
+  return ((stack1 = container.invokePartial(lookupProperty(partials,"zodSchema"),lookupProperty(depth0,"link"),{"name":"zodSchema","data":data,"helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "");
+},"3":function(container,depth0,helpers,partials,data) {
     var stack1, lookupProperty = container.lookupProperty || function(parent, propertyName) {
         if (Object.prototype.hasOwnProperty.call(parent, propertyName)) {
           return parent[propertyName];
@@ -21,37 +21,20 @@ export default {"1":function(container,depth0,helpers,partials,data) {
         return undefined
     };
 
-  return ((stack1 = container.lambda(container.strict(lookupProperty(depth0,"link"), "alias", {"start":{"line":1,"column":41},"end":{"line":1,"column":51}} ), depth0)) != null ? stack1 : "")
-    + "Schema";
-},"4":function(container,depth0,helpers,partials,data) {
-    var stack1, lookupProperty = container.lookupProperty || function(parent, propertyName) {
-        if (Object.prototype.hasOwnProperty.call(parent, propertyName)) {
-          return parent[propertyName];
-        }
-        return undefined
-    };
-
-  return ((stack1 = container.lambda(container.strict(lookupProperty(depth0,"link"), "type", {"start":{"line":1,"column":71},"end":{"line":1,"column":80}} ), depth0)) != null ? stack1 : "")
-    + "Schema";
-},"6":function(container,depth0,helpers,partials,data) {
-    var stack1;
-
-  return "z."
-    + ((stack1 = container.lambda(container.strict(depth0, "base", {"start":{"line":1,"column":109},"end":{"line":1,"column":113}} ), depth0)) != null ? stack1 : "")
-    + "()";
-},"8":function(container,depth0,helpers,partials,data) {
+  return ((stack1 = lookupProperty(helpers,"zodBaseSchema").call(depth0 != null ? depth0 : (container.nullContext || {}),lookupProperty(depth0,"base"),{"name":"zodBaseSchema","hash":{},"data":data,"loc":{"start":{"line":1,"column":47},"end":{"line":1,"column":69}}})) != null ? stack1 : "");
+},"5":function(container,depth0,helpers,partials,data) {
     var stack1;
 
   return ".min("
-    + ((stack1 = container.lambda(container.strict(depth0, "minItems", {"start":{"line":1,"column":150},"end":{"line":1,"column":158}} ), depth0)) != null ? stack1 : "")
+    + ((stack1 = container.lambda(container.strict(depth0, "minItems", {"start":{"line":1,"column":101},"end":{"line":1,"column":109}} ), depth0)) != null ? stack1 : "")
     + ")";
-},"10":function(container,depth0,helpers,partials,data) {
+},"7":function(container,depth0,helpers,partials,data) {
     var stack1;
 
   return ".max("
-    + ((stack1 = container.lambda(container.strict(depth0, "maxItems", {"start":{"line":1,"column":193},"end":{"line":1,"column":201}} ), depth0)) != null ? stack1 : "")
+    + ((stack1 = container.lambda(container.strict(depth0, "maxItems", {"start":{"line":1,"column":144},"end":{"line":1,"column":152}} ), depth0)) != null ? stack1 : "")
     + ")";
-},"12":function(container,depth0,helpers,partials,data) {
+},"9":function(container,depth0,helpers,partials,data) {
     return ".nullable()";
 },"compiler":[8,">= 4.3.0"],"main":function(container,depth0,helpers,partials,data) {
     var stack1, alias1=depth0 != null ? depth0 : (container.nullContext || {}), lookupProperty = container.lookupProperty || function(parent, propertyName) {
@@ -62,9 +45,10 @@ export default {"1":function(container,depth0,helpers,partials,data) {
     };
 
   return "z.array("
-    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"link"),{"name":"if","hash":{},"fn":container.program(1, data, 0),"inverse":container.program(6, data, 0),"data":data,"loc":{"start":{"line":1,"column":8},"end":{"line":1,"column":125}}})) != null ? stack1 : "")
+    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"link"),{"name":"if","hash":{},"fn":container.program(1, data, 0),"inverse":container.program(3, data, 0),"data":data,"loc":{"start":{"line":1,"column":8},"end":{"line":1,"column":76}}})) != null ? stack1 : "")
     + ")"
-    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"minItems"),{"name":"if","hash":{},"fn":container.program(8, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":1,"column":126},"end":{"line":1,"column":169}}})) != null ? stack1 : "")
-    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"maxItems"),{"name":"if","hash":{},"fn":container.program(10, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":1,"column":169},"end":{"line":1,"column":212}}})) != null ? stack1 : "")
-    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"isNullable"),{"name":"if","hash":{},"fn":container.program(12, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":1,"column":212},"end":{"line":1,"column":248}}})) != null ? stack1 : "");
-},"useData":true}
+    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"minItems"),{"name":"if","hash":{},"fn":container.program(5, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":1,"column":77},"end":{"line":1,"column":120}}})) != null ? stack1 : "")
+    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"maxItems"),{"name":"if","hash":{},"fn":container.program(7, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":1,"column":120},"end":{"line":1,"column":163}}})) != null ? stack1 : "")
+    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"isNullable"),{"name":"if","hash":{},"fn":container.program(9, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":1,"column":163},"end":{"line":1,"column":199}}})) != null ? stack1 : "")
+    + "\n";
+},"usePartial":true,"useData":true}