+++ src/templatesCompiled/client/zod/partials/zodSchemaDictionary.ts
@@ -12,8 +12,8 @@ export default {"1":function(container,depth0,helpers,partials,data) {
         return undefined
     };
 
-  return ((stack1 = lookupProperty(helpers,"if").call(depth0 != null ? depth0 : (container.nullContext || {}),lookupProperty(lookupProperty(depth0,"link"),"alias"),{"name":"if","hash":{},"fn":container.program(2, data, 0),"inverse":container.program(4, data, 0),"data":data,"loc":{"start":{"line":1,"column":21},"end":{"line":1,"column":97}}})) != null ? stack1 : "");
-},"2":function(container,depth0,helpers,partials,data) {
+  return ((stack1 = container.invokePartial(lookupProperty(partials,"zodSchema"),lookupProperty(depth0,"link"),{"name":"zodSchema","data":data,"helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "");
+},"3":function(container,depth0,helpers,partials,data) {
     var stack1, lookupProperty = container.lookupProperty || function(parent, propertyName) {
         if (Object.prototype.hasOwnProperty.call(parent, propertyName)) {
           return parent[propertyName];
@@ -21,25 +21,8 @@ export default {"1":function(container,depth0,helpers,partials,data) {
         return undefined
     };
 
-  return ((stack1 = container.lambda(container.strict(lookupProperty(depth0,"link"), "alias", {"start":{"line":1,"column":42},"end":{"line":1,"column":52}} ), depth0)) != null ? stack1 : "")
-    + "Schema";
-},"4":function(container,depth0,helpers,partials,data) {
-    var stack1, lookupProperty = container.lookupProperty || function(parent, propertyName) {
-        if (Object.prototype.hasOwnProperty.call(parent, propertyName)) {
-          return parent[propertyName];
-        }
-        return undefined
-    };
-
-  return ((stack1 = container.lambda(container.strict(lookupProperty(depth0,"link"), "type", {"start":{"line":1,"column":72},"end":{"line":1,"column":81}} ), depth0)) != null ? stack1 : "")
-    + "Schema";
-},"6":function(container,depth0,helpers,partials,data) {
-    var stack1;
-
-  return "z."
-    + ((stack1 = container.lambda(container.strict(depth0, "base", {"start":{"line":1,"column":110},"end":{"line":1,"column":114}} ), depth0)) != null ? stack1 : "")
-    + "()";
-},"8":function(container,depth0,helpers,partials,data) {
+  return ((stack1 = lookupProperty(helpers,"zodBaseSchema").call(depth0 != null ? depth0 : (container.nullContext || {}),lookupProperty(depth0,"base"),{"name":"zodBaseSchema","hash":{},"data":data,"loc":{"start":{"line":1,"column":60},"end":{"line":1,"column":82}}})) != null ? stack1 : "");
+},"5":function(container,depth0,helpers,partials,data) {
     return ".nullable()";
 },"compiler":[8,">= 4.3.0"],"main":function(container,depth0,helpers,partials,data) {
     var stack1, alias1=depth0 != null ? depth0 : (container.nullContext || {}), lookupProperty = container.lookupProperty || function(parent, propertyName) {
@@ -49,8 +32,9 @@ export default {"1":function(container,depth0,helpers,partials,data) {
         return undefined
     };
 
-  return "z.record("
-    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"link"),{"name":"if","hash":{},"fn":container.program(1, data, 0),"inverse":container.program(6, data, 0),"data":data,"loc":{"start":{"line":1,"column":9},"end":{"line":1,"column":126}}})) != null ? stack1 : "")
+  return "z.record(z.string(), "
+    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"link"),{"name":"if","hash":{},"fn":container.program(1, data, 0),"inverse":container.program(3, data, 0),"data":data,"loc":{"start":{"line":1,"column":21},"end":{"line":1,"column":89}}})) != null ? stack1 : "")
     + ")"
-    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"isNullable"),{"name":"if","hash":{},"fn":container.program(8, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":1,"column":127},"end":{"line":1,"column":163}}})) != null ? stack1 : "");
-},"useData":true}
+    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(depth0,"isNullable"),{"name":"if","hash":{},"fn":container.program(5, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":1,"column":90},"end":{"line":1,"column":126}}})) != null ? stack1 : "")
+    + "\n";
+},"usePartial":true,"useData":true}